feat: Make app speak text when you tap in a bounding box with text
Here is the URL of the codelab you are taking.  This will come in pretty handy
for barcode stuff and when you need to read in recipes.  You can use the text in
the bounding boxes.  But you will have to have an easy way to correct the text.
You should see if there is something on github for that.
https://codelabs.developers.google.com/codelabs/mobile-vision-ocr/